:green_heart: Fix tests after pytest/tox updates
The plugin pytest-django started doing weird stuff, which we notice
because of doing raw database stuff. Should be good again now for a
while.

The database dumps have been updated to no longer include the database
creation, because the restore code now takes care of dropping and
creating the target database.